pyproject.toml
[project]
name = "lingma-xiaoshuo-assistant"
version = "0.1.16"
description = "灵码小说助手"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.121.2",
    "pydantic>=2.12.4",
    "pydantic-settings>=2.12.0",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "uvicorn[standard]>=0.38.0",
    "tortoise-orm[asyncpg]>=0.21.7",
    "aerich>=0.7.2",
    "aiosqlite>=0.20.0",
    "loguru>=0.7.3",
    "bcrypt>=5.0.0",
    "psutil>=7.1.3",
    "sse-starlette>=3.0.3",
    "openai>=2.8.1",
]

[tool.aerich]
tortoise_orm = "src.backend.config.database.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."


[tool.ruff]
extend-exclude = []
ignore = [
    "E402",
    "E501",
    "E722",
    "B008",
    "B905",
    "F401",
    "TID252",
    "TRY002",
    "TRY003",
    "RUF001",
    "RUF002",
    "RUF003",
    "RUF006",
    "RUF012",
    "RUF100",
    "PGH003",
    "N801",
    "A003",
]
select = [
    "F",
    "E",
    "I",
    "B",
    "A",
    "COM",
    "C4",
    "ISC",
    "PIE",
    "PYI",
    "Q",
    "RSE",
    "RET",
    "SLF",
    "SIM",
    "TID",
    "ARG",
    "PTH",
    "PGH",
    "TRY",
    "RUF",
]

[dependency-groups]
dev = ["pyinstaller>=6.16.0"]
